REFACT: move vprint to utils
